Add more option for nginx config

+++ bin/start.sh
@@ -11,10 +11,12 @@ if [[ -z "${ENV_SERVER_NAME}" ]]; then
 else
   SERVER_NAME="${ENV_SERVER_NAME}"
 fi
-cp /etc/nginx/conf.d/site.conf.tpl /etc/nginx/conf.d/site.conf
+if [ ! -f /etc/nginx/conf.d/site.conf ]; then
+   	cp /etc/nginx/conf.d/site.conf.tpl /etc/nginx/conf.d/site.conf
+	sed -i "s|SERVER_NAME|$SERVER_NAME|g" /etc/nginx/conf.d/site.conf
+	sed -i "s|ROOT_PATH|$ROOT_PATH|g" /etc/nginx/conf.d/site.conf
+fi
 
-sed -i "s|SERVER_NAME|$SERVER_NAME|g" /etc/nginx/conf.d/site.conf
-sed -i "s|ROOT_PATH|$ROOT_PATH|g" /etc/nginx/conf.d/site.conf
 
 if [ -z "$(ls -A ${ROOT_PATH})" ]; then
    echo "<?=phpinfo()?>" > "${ROOT_PATH}/index.php"
